@@ -20,13 +20,13 @@
     "history": "^3.0.0",
     "js-cookie": "^2.1.4",
     "lodash": "^4.17.10",
-    "material-ui": "^0.18.7",
+    "material-ui": "^0.20.0",
     "moment": "^2.22.1",
     "moment-timezone": "^0.5.16",
-    "react": "^15.4.0",
-    "react-dom": "^15.4.0",
-    "react-mounter": "1.2.0",
-    "react-router-dom": "^4.0.0-beta.6",
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2",
+    "react-mounter": "^1.2.0",
+    "react-router-dom": "^4.2.2",
     "react-tap-event-plugin": "^2.0.1",
     "suncalc": "^1.8.0",
     "xmldom": "^0.1.27",
