@@ -1,6 +1,4 @@
 import React from 'react';
-import {Controller} from '../domains/controller.js';
-import {Colors} from '../../lib/colors.js';
 import FlatButton from 'material-ui/FlatButton';
 import Dialog from 'material-ui/Dialog';
 import Toggle from 'material-ui/Toggle';
@@ -8,74 +6,64 @@ import SelectField from 'material-ui/SelectField';
 import MenuItem from 'material-ui/MenuItem';
 import Checkbox from 'material-ui/Checkbox';
 import Slider from 'material-ui/Slider';
+import {Controller} from '../domains/controller.js';
+import {Colors} from '../../lib/colors.js';
+import {withTracker} from 'meteor/react-meteor-data';
 
 
-export const SettingsMenu = React.createClass({
-
-	mixins: [ReactMeteorData],
-	getMeteorData: function() {
-		return {
-			open: Controller.settingsOpen(),
-			clockSettings: Controller.getClockSettings(),
-			availableUnitModes: Controller.getAvailableUnitModes(),
-			availableClockSizes: Controller.getAvailableClockSizes(),
-			colorTheme: Controller.getColorTheme(),
-			colorThemes: Colors.getColorThemes(),
-			hue: Controller.getHue()
-		};
-	},
+class Settings extends React.Component {
 
 	handleDialogClose() {
 		Dispatcher.dispatch({
 			actionType: "DOT_MENU_CLOSE_BUTTON_CLICKED"
 		});
-	},
+	}
 
 	tzModeToggled() {
 		Dispatcher.dispatch({
 			actionType: "FORECAST_TIMEZONE_TOGGLED"
 		});
-	},
+	}
 
 	gradientModeToggled() {
 		Dispatcher.dispatch({
 			actionType: "GRADIENT_MODE_TOGGLED"
 		});
-	},
+	}
 
 	secondHandToggled() {
 		Dispatcher.dispatch({
 			actionType: "CLOCK_SECOND_HAND_TOGGLED"
 		});
-	},
+	}
 
 	colorThemeChange(event, index, value) {
 		Dispatcher.dispatch({
 			actionType: "COLOR_THEME_SELECTED",
 			data: value
 		});
-	},
+	}
 
 	hueChange(event, value) {
 		Dispatcher.dispatch({
 			actionType: "HUE_CHANGED",
 			data: value
 		});
-	},
+	}
 
 	unitModeChange(event, index, value) {
 		Dispatcher.dispatch({
 			actionType: "UNIT_MODE_SELECTED",
 			data: value
 		});
-	},
+	}
 
 	clockSizeChange(event, index, value) {
 		Dispatcher.dispatch({
 			actionType: "CLOCK_SIZE_SELECTED",
 			data: value
 		});
-	},
+	}
 
 	render() {
 
@@ -94,7 +82,7 @@ export const SettingsMenu = React.createClass({
 			/>
 		];
 
-		const colorThemeItems = this.data.colorThemes.map(function(item) {
+		const colorThemeItems = this.props.colorThemes.map(function(item) {
 			return (
 				<MenuItem
 					primaryText={item.name}
@@ -104,7 +92,7 @@ export const SettingsMenu = React.createClass({
 			);
 		});
 
-		const unitModeItems = this.data.availableUnitModes.map(function(item) {
+		const unitModeItems = this.props.availableUnitModes.map(function(item) {
 			return (
 				<MenuItem
 					primaryText={item.text}
@@ -114,7 +102,7 @@ export const SettingsMenu = React.createClass({
 			);
 		});
 
-		const clockSizeItems = this.data.availableClockSizes.map(function(item) {
+		const clockSizeItems = this.props.availableClockSizes.map(function(item) {
 			return (
 				<MenuItem
 					primaryText={item.text}
@@ -129,24 +117,24 @@ export const SettingsMenu = React.createClass({
 				<div className="settings-check-boxes">
 					<Checkbox
 						label="Gradient mode"
-						checked={this.data.clockSettings.gradientMode}
+						checked={this.props.clockSettings.gradientMode}
 						onCheck={this.gradientModeToggled}
 					/>
 					<Checkbox
 						label="Forecast timezone"
-						checked={this.data.clockSettings.forecastTimezone}
+						checked={this.props.clockSettings.forecastTimezone}
 						onCheck={this.tzModeToggled}
 					/>
 					<Checkbox
 						label="Second hand"
-						checked={this.data.clockSettings.secondHand}
+						checked={this.props.clockSettings.secondHand}
 						onCheck={this.secondHandToggled}
 					/>
 				</div>
 				<div className="settings-select-fields">
 					<SelectField
 						floatingLabelText="Color theme"
-						value={this.data.colorTheme.id}
+						value={this.props.colorTheme.id}
 						onChange={this.colorThemeChange}
 						style={selectFieldStyle}
 					>
@@ -159,13 +147,13 @@ export const SettingsMenu = React.createClass({
 					<div className="settings-slider-container">
 						<Slider
 							step={0.05}
-							value={this.data.hue}
+							value={this.props.hue}
 							onChange={this.hueChange}
 						/>
 					</div>
 					<SelectField
 						floatingLabelText="Units"
-						value={this.data.clockSettings.unitMode.id}
+						value={this.props.clockSettings.unitMode.id}
 						onChange={this.unitModeChange}
 						style={selectFieldStyle}
 					>
@@ -174,7 +162,7 @@ export const SettingsMenu = React.createClass({
 					<br />
 					<SelectField
 						floatingLabelText="Clock size"
-						value={this.data.clockSettings.clockSize.id}
+						value={this.props.clockSettings.clockSize.id}
 						onChange={this.clockSizeChange}
 						style={selectFieldStyle}
 					>
@@ -191,7 +179,7 @@ export const SettingsMenu = React.createClass({
 					actions={actions}
 					modal={false}
 					onRequestClose={this.handleDialogClose}
-					open={this.data.open}
+					open={this.props.open}
 					contentStyle={dialogStyle}
 					autoScrollBodyContent={true}
 				>
@@ -200,4 +188,16 @@ export const SettingsMenu = React.createClass({
 			</div>
 		);
 	}
-});
+}
+
+export default SettingsContainer = withTracker(props => {
+	return {
+		open: Controller.settingsOpen(),
+		clockSettings: Controller.getClockSettings(),
+		availableUnitModes: Controller.getAvailableUnitModes(),
+		availableClockSizes: Controller.getAvailableClockSizes(),
+		colorTheme: Controller.getColorTheme(),
+		colorThemes: Colors.getColorThemes(),
+		hue: Controller.getHue()
+	};
+})(Settings);
